inner/cicle_6/ft_transcendence/backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyotp==2.9.0
redis==5.2.1
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
Pillow==10.2.0
gunicorn>=20.0.0
django-cors-headers # cors 세팅